--- stable_baselines/a2c/utils.py
@@ -1,15 +1,0 @@
-def discount_with_dones(rewards, dones, gamma):
-    """
-    Apply the discount value to the reward, where the environment is not done
-
-    :param rewards: ([float]) The rewards
-    :param dones: ([bool]) Whether an environment is done or not
-    :param gamma: (float) The discount value
-    :return: ([float]) The discounted rewards
-    """
-    discounted = []
-    ret = 0  # Return: discounted reward
-    for reward, done in zip(rewards[::-1], dones[::-1]):
-        ret = reward + gamma * ret * (1. - done)  # fixed off by one bug
-        discounted.append(ret)
-    return discounted[::-1]
